+++ pom.xml
@@ -2,12 +2,12 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.tyson-spec.tyson</groupId>
-  <artifactId>TysonInstance</artifactId>
+  <groupId>org.rumbledb.tyson</groupId>
+  <artifactId>jtyson</artifactId>
   <version>0.0.1</version>
   <packaging>jar</packaging>
 
-  <name>TysonInstance</name>
+  <name>jtyson</name>
   <url>http://maven.apache.org</url>
 
   <properties>
