+++ src/main/java/org/rumbledb/tyson/TysonFactory.java
@@ -1,4 +1,4 @@
-package org.tyson_spec.tyson.TysonInstance;
+package org.rumbledb.tyson;
 
 import java.math.BigDecimal;
