+++ src/main/java/org/rumbledb/tyson/TysonObject.java
@@ -1,4 +1,4 @@
-package org.tyson_spec.tyson.TysonInstance;
+package org.rumbledb.tyson;
 
 import java.util.List;
 import java.util.ArrayList;
