+++ src/main/java/org/rumbledb/tyson/example_tyson.txt
