+++ src/main/java/org/rumbledb/tyson/tyson_types.txt
