+++ src/test/java/org/rumbledb/tyson/tyson_types.txt
