+++ src/test/java/org/rumbledb/tyson/zoo_example.txt
